dtype added to pimages

+++ supervised_learning/object_detection/5-yolo.py
@@ -67,7 +67,7 @@ class Yolo:
 
             pimages.append(preprocessed)
 
-        pimages = np.array(pimages)
+        pimages = np.array(pimages, dtype=np.float32)
         image_shapes = np.array(image_shapes)
 
         return (pimages, image_shapes)
